backup/gynecology_chatbot_backup_20250519/chainlit_app/requirements.txt
# Chainlit and related packages
chainlit==1.0.101
aiohttp==3.9.1
python-dotenv==1.0.0

# LLM APIs (for direct use if needed)
openai==1.3.7
google-generativeai==0.3.1
anthropic==0.5.0

# Utilities
pydantic==2.4.2
